# docs/swagger.yaml
definitions:
  errs.Err:
    properties:
      message:
        type: string
    type: object
  wallet.Request:
    properties:
      balance:
        example: 100
        type: number
      user_id:
        example: 1
        type: integer
      user_name:
        example: John Doe
        type: string
      wallet_name:
        example: John's Wallet
        type: string
      wallet_type:
        enum:
        - Savings
        - Credit Card
        - Crypto Wallet
        example: Credit Card
        type: string
    required:
    - balance
    - user_id
    - user_name
    - wallet_name
    - wallet_type
    type: object
  wallet.Wallet:
    properties:
      balance:
        example: 100
        type: number
      created_at:
        example: "2024-03-25T14:19:00.729237Z"
        type: string
      id:
        example: 1
        type: integer
      user_id:
        example: 1
        type: integer
      user_name:
        example: John Doe
        type: string
      wallet_name:
        example: John's Wallet
        type: string
      wallet_type:
        example: Create Card
        type: string
    type: object
host: localhost:1323
info:
  contact: {}
  description: Sophisticated Wallet API
  title: Wallet API
  version: "1.0"
paths:
  /api/v1/users/{id}/wallets:
    get:
      consumes:
      - application/json
      description: Get user wallets
      parameters:
      - description: user id
        in: path
        name: id
        required: true
        type: string
      - description: wallet type
        enum:
        - Savings
        - Credit Card
        - Crypto Wallet
        in: query
        name: wallet_type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/wallet.Wallet'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errs.Err'
      summary: Get user wallets
      tags:
      - wallet
  /api/v1/wallets:
    get:
      consumes:
      - application/json
      description: Get all wallets
      parameters:
      - description: wallet type
        enum:
        - Savings
        - Credit Card
        - Crypto Wallet
        in: query
        name: wallet_type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/wallet.Wallet'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errs.Err'
      summary: Get all wallets
      tags:
      - wallet
    post:
      consumes:
      - application/json
      description: Create wallet
      parameters:
      - description: wallet request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/wallet.Request'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/wallet.Wallet'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errs.Err'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errs.Err'
      summary: Create wallet
      tags:
      - wallet
  /api/v1/wallets/{id}:
    delete:
      consumes:
      - application/json
      description: Delete wallet
      parameters:
      - description: wallet id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errs.Err'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errs.Err'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errs.Err'
      summary: Delete wallet
      tags:
      - wallet
    put:
      consumes:
      - application/json
      description: Update wallet
      parameters:
      - description: wallet id
        in: path
        name: id
        required: true
        type: string
      - description: wallet request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/wallet.Request'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errs.Err'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errs.Err'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errs.Err'
      summary: Update wallet
      tags:
      - wallet
swagger: "2.0"